requirements.txt
gevent
grequests
bottle
nose-gevented-multiprocess
